Ensure onStdinClose only completes for non-tty streams
The initial stdin commit 144cd35 inverted the TTY conditional
b/w the io/node and io/vm implementations of onStdinClose.
The node implementation incorrectly checked for the presence
of TTY to complete the stream, while the vm implementation
correctly checked for its absence. The commit that landed
upstream c7ab426 normalized the incorrect behaviour, which
means that sass was still not closing when stdin was closed,
unless stdin was a TTY.

Unfortunately that created a "worst of both worlds" situation
because programs that start sass and then close unexpectedly
will still leave zombie sass processes running in the
background, and wrapper scripts designed to mitigate this
exact problem will stop working because moving the process to
the background now incorrectly causes the job to stop.

This change ensures we only complete the CancelableOperation
onStdinClose for non-tty standard input streams.