Fix for case mentioned by @nex3 in issue #2601
..in comment: https://github.com/sass/dart-sass/issues/2601#issuecomment-3046528111

This solution uses a "lazy algorithm" for catching the error. Note that the diff makes things look more drastic than they are: I simply moved the entire if-then logic for dealing with already-loaded modules.

A better method might be to write a dedicated function to traverse the node tree with fewer potential side-effects.(I started on this but eventually set it aside as being more complicated than the current solution.)