Parse selectors in the sass-parser package (#2670)
This adds a separate "interpolated selector" AST that parallels the
existing selector AST, but includes the interpolation that's present
in the source files. This AST isn't used by the Sass implementation
itself; it's just made available so that tools can gracefully interact
with selectors as an AST in unevaluated Sass files.